# ingestor/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ingestor (Karma8 Test Case)",
        "title": "API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:7788",
    "basePath": "/",
    "paths": {
        "/ingestor/file/download": {
            "post": {
                "description": "Download file by bucket name and key value",
                "produces": [
                    "application/octet-stream"
                ],
                "summary": "Download file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bucket name for target file",
                        "name": "X-Karma8-Object-Bucket",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key value for target file",
                        "name": "X-Karma8-Object-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/ingestor/file/upload": {
            "post": {
                "description": "Upload file with bucket name and key value",
                "consumes": [
                    "application/octet-stream"
                ],
                "summary": "Upload file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bucket name for target file",
                        "name": "X-Karma8-Object-Bucket",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key value for target file",
                        "name": "X-Karma8-Object-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Total size of target file",
                        "name": "X-Karma8-Object-Total-Size",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Target file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}